# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOC Triage Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/logs:
    get:
      operationId: getSessionLogs
      tags: [logs]
      summary: Get logs for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    post:
      operationId: addLogToSession
      tags: [logs]
      summary: Add a log entry to a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogEntryInput"
      responses:
        "201":
          description: Log entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"
  /sessions/{id}/logs/{logId}:
    delete:
      operationId: removeLogFromSession
      tags: [logs]
      summary: Remove a log entry from session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /sessions/{id}/analyze:
    post:
      operationId: analyzeSession
      tags: [sessions]
      summary: Trigger AI analysis via n8n SOAR webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Bad request
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LogEntry:
      type: object
      required: [id, sessionId, source, rawJson, extractedIp, masked, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        source:
          type: string
          enum: [fortigate, watchguard, agent_windows, agent_linux, unknown]
        rawJson:
          type: string
        extractedIp:
          type: ["string", "null"]
        masked:
          type: boolean
        createdAt:
          type: string
    LogEntryInput:
      type: object
      required: [source, rawJson]
      properties:
        source:
          type: string
          enum: [fortigate, watchguard, agent_windows, agent_linux, unknown]
        rawJson:
          type: string
...
    AnalyzeInput:
      type: object
      required: [maskIps]
      properties:
        maskIps:
          type: boolean
        additionalContext:
          type: string
    Report:
      type: object
      required: [id, sessionId, summary, severity, iocs, recommendations, rawAiResponse, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        summary:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, informational]
        iocs:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        attackVector:
          type: ["string", "null"]
        affectedSystems:
          type: array
          items:
            type: string
        rawAiResponse:
          type: string
        n8nExecutionId:
          type: ["string", "null"]
        createdAt:
          type: string
...