# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOC Triage Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  maskIps: boolean;
  additionalContext?: string;
}