# lib/api-zod/src/generated/types/logEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOC Triage Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntryInputSource } from './logEntryInputSource';

export interface LogEntryInput {
  source: LogEntryInputSource;
  rawJson: string;
}